+++ ibm/service/configurationaggregator/README.md
@@ -4,5 +4,8 @@ This area is primarily for IBM provider contributors and maintainers. For inform
 
 
 ## Handy Links
+* [Find out about contributing](../../../CONTRIBUTING.md) to the IBM provider!
+* IBM Provider Docs: [Home](https://registry.terraform.io/providers/IBM-Cloud/ibm/latest/docs)
+* IBM Provider Docs: [One of the  resources](https://registry.terraform.io/providers/IBM-Cloud/ibm/latest/docs/resources/config_aggregator_settings)
 * IBM API Docs: [IBM API Docs for ]()
-* IBM  SDK: [IBM SDK for ](https://github.com/IBM/configuration-aggregator-go-sdk/tree/main/configurationaggregatorv1")
+* IBM  SDK: [IBM SDK for ](https://github.com/IBM/appconfiguration-go-admin-sdk/tree/master/configurationaggregatorv1)
