+++ ibm/service/configurationaggregator/data_source_ibm_config_aggregator_settings.go
@@ -16,10 +16,9 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	"github.com/IBM/configuration-aggregator-go-sdk/configurationaggregatorv1"
-
 	"github.com/IBM-Cloud/terraform-provider-ibm/ibm/conns"
 	"github.com/IBM-Cloud/terraform-provider-ibm/ibm/flex"
+	"github.com/IBM/configuration-aggregator-go-sdk/configurationaggregatorv1"
 )
 
 func DataSourceIbmConfigAggregatorSettings() *schema.Resource {
